+++ api-contract/Abi/toLatestCompatible.ts
@@ -1,5 +1,5 @@
 
-import type { ContractMetadataV4, ContractMetadataV5 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { ContractMetadataV4, ContractMetadataV5, ContractMetadataV6 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { Registry } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { ContractMetadataSupported } from './index.ts';
 
@@ -8,7 +8,7 @@ import { v1ToV2 } from './toV2.ts';
 import { v2ToV3 } from './toV3.ts';
 import { v3ToV4 } from './toV4.ts';
 
-export const enumVersions = ['V5', 'V4', 'V3', 'V2', 'V1'] as const;
+export const enumVersions = ['V6', 'V5', 'V4', 'V3', 'V2', 'V1'] as const;
 
 type Versions = typeof enumVersions[number] | 'V0';
 
@@ -17,6 +17,10 @@ type Converter = (registry: Registry, vx: any) => ContractMetadataSupported;
 function createConverter <I, O> (next: (registry: Registry, input: O) => ContractMetadataSupported, step: (registry: Registry, input: I) => O): (registry: Registry, input: I) => ContractMetadataSupported {
   return (registry: Registry, input: I): ContractMetadataSupported =>
     next(registry, step(registry, input));
+}
+
+export function v6ToLatestCompatible (_registry: Registry, v6: ContractMetadataV6): ContractMetadataV6 {
+  return v6;
 }
 
 export function v5ToLatestCompatible (_registry: Registry, v5: ContractMetadataV5): ContractMetadataV5 {
@@ -33,6 +37,7 @@ export const v1ToLatestCompatible = /*#__PURE__*/ createConverter(v2ToLatestComp
 export const v0ToLatestCompatible = /*#__PURE__*/ createConverter(v1ToLatestCompatible, v0ToV1);
 
 export const convertVersions: [Versions, Converter][] = [
+  ['V6', v6ToLatestCompatible],
   ['V5', v5ToLatestCompatible],
   ['V4', v4ToLatestCompatible],
   ['V3', v3ToLatestCompatible],