+++ api-contract/base/Contract.ts
@@ -2,7 +2,7 @@
 import type { ApiBase } from 'https://deno.land/x/polkadot/api/base/index.ts';
 import type { SubmittableExtrinsic } from 'https://deno.land/x/polkadot/api/submittable/types.ts';
 import type { ApiTypes, DecorateMethod } from 'https://deno.land/x/polkadot/api/types/index.ts';
-import type { AccountId, ContractExecResult, EventRecord, Weight, WeightV2 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { AccountId, AccountId20, ContractExecResult, EventRecord, Weight, WeightV2 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { ISubmittableResult } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { Abi } from '../Abi/index.ts';
 import type { AbiMessage, ContractCallOutcome, ContractOptions, DecodedEvent, WeightAll } from '../types.ts';
@@ -49,15 +49,15 @@ export class Contract<ApiType extends ApiTypes> extends Base<ApiType> {
   /**
    * @description The on-chain address for this contract
    */
-  readonly address: AccountId;
+  readonly address: AccountId | AccountId20;
 
   readonly #query: MapMessageQuery<ApiType> = {};
   readonly #tx: MapMessageTx<ApiType> = {};
 
-  constructor (api: ApiBase<ApiType>, abi: string | Record<string, unknown> | Abi, address: string | AccountId, decorateMethod: DecorateMethod<ApiType>) {
+  constructor (api: ApiBase<ApiType>, abi: string | Record<string, unknown> | Abi, address: string | AccountId | AccountId20, decorateMethod: DecorateMethod<ApiType>) {
     super(api, abi, decorateMethod);
 
-    this.address = this.registry.createType('AccountId', address);
+    this.address = this.registry.createType(this._isRevive ? 'AccountId20' : 'AccountId', address);
 
     this.abi.messages.forEach((m): void => {
       if (isUndefined(this.#tx[m.method])) {
@@ -97,7 +97,9 @@ export class Contract<ApiType extends ApiTypes> extends Base<ApiType> {
   };
 
   #exec = (messageOrId: AbiMessage | string | number, { gasLimit = BN_ZERO, storageDepositLimit = null, value = BN_ZERO }: ContractOptions, params: unknown[]): SubmittableExtrinsic<ApiType> => {
-    return this.api.tx.contracts.call(
+    const palletTx = this._isRevive ? this.api.tx.revive : this.api.tx.contracts;
+
+    return palletTx.call(
       this.address,
       value,
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -131,7 +133,9 @@ export class Contract<ApiType extends ApiTypes> extends Base<ApiType> {
     return {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
       send: this._decorateMethod((origin: string | AccountId | Uint8Array) =>
-        this.api.rx.call.contractsApi.call<ContractExecResult>(
+        (this._isRevive
+          ? this.api.rx.call.reviveApi.call
+          : this.api.rx.call.contractsApi.call)<ContractExecResult>(
           origin,
           this.address,
           value,