+++ api-contract/test/contracts/ink/v6/index.ts
@@ -1,0 +1,3 @@
+
+export { default as erc20Contract } from './erc20.contract.json' assert { type: 'json' };
+export { default as erc20Metadata } from './erc20.json' assert { type: 'json' };