+++ types/interfaces/contractsAbi/definitions.ts
@@ -124,6 +124,13 @@ const spec = {
     docs: 'Vec<Text>',
     environment: 'ContractEnvironmentV4'
   },
+  ContractContractSpecV6: {
+    constructors: 'Vec<ContractConstructorSpecV4>',
+    messages: 'Vec<ContractMessageSpecV3>',
+    events: 'Vec<ContractEventSpecV3>',
+    docs: 'Vec<Text>',
+    environment: 'ContractEnvironmentV4'
+  },
 
   ContractDisplayName: 'SiPath',
   ContractEventParamSpecV0: {
@@ -218,7 +225,7 @@ const latest = {
   ContractEventParamSpecLatest: 'ContractEventParamSpecV2',
   ContractMessageParamSpecLatest: 'ContractMessageParamSpecV2',
   ContractMessageSpecLatest: 'ContractMessageSpecV3',
-  ContractMetadataLatest: 'ContractMetadataV5'
+  ContractMetadataLatest: 'ContractMetadataV6'
 };
 
 export default {
@@ -229,6 +236,10 @@ export default {
     ...latest,
     ContractProjectInfo: {
       source: 'ContractProjectSource',
+      contract: 'ContractProjectContract'
+    },
+    ContractReviveProjectInfo: {
+      source: 'ContractReviveProjectSource',
       contract: 'ContractProjectContract'
     },
     ContractMetadataV0: {
@@ -258,6 +269,11 @@ export default {
       spec: 'ContractContractSpecV5',
       version: 'u64'
     },
+    ContractMetadataV6: {
+      types: 'Vec<PortableType>',
+      spec: 'ContractContractSpecV6',
+      version: 'u64'
+    },
     ContractMetadata: {
       _enum: {
         V0: 'ContractMetadataV0',
@@ -265,7 +281,8 @@ export default {
         V2: 'ContractMetadataV2',
         V3: 'ContractMetadataV3',
         V4: 'ContractMetadataV4',
-        V5: 'ContractMetadataV5'
+        V5: 'ContractMetadataV5',
+        V6: 'ContractMetadataV6'
       }
     },
     ContractProjectV0: {
@@ -292,6 +309,16 @@ export default {
     ContractProjectSource: {
       _alias: {
         wasmHash: 'hash'
+      },
+      wasmHash: '[u8; 32]',
+      language: 'Text',
+      compiler: 'Text',
+      wasm: 'Raw'
+    },
+    ContractReviveProjectSource: {
+      _alias: {
+        wasmHash: 'hash',
+        wasm: 'contract_binary'
       },
       wasmHash: '[u8; 32]',
       language: 'Text',